Fix a bug in the attachment handling procedure.
When comparing a new file to an existing one, the first check is done on
the filename.

ELOG is replacing all spaces with _ internally. So if you upload a new
attachment with spaces in the filename, this will never match any
existing attachment simply because spaces are removed.

In order not to limit the freedom of the user, the filename of the new
attachments is manipulated internally to replace spaces with _.